Fix compiler warnings
I changed the warning flags for the compiler to have warnings cause
errors and for there to be every compiler warning known to gcc.  I then
went about fixing the compiler warnings.  The two biggest offenders were
-Wsign-compare and -Wunused-parameter.

Fixing these did involve changing some function handles to have the
`unsigned int` type in a few places.  I noticed that most instances of
these `unsigned int`s were for representing coordinates, so I think a
refactor to create a type `distance` that is a type alias for
`unsigned int` or similar could be useful.

This was also a very quick cleanup to make the compiler happy.  There
are spots that could be cleaned a lot further.